refactor: 💡 remaining integration tests, split helpers up
