+++ src/index.ts
@@ -132,7 +132,7 @@ class ForkTsCheckerWebpackPlugin {
   private tslintVersion: string;
 
   private service?: childProcess.ChildProcess;
-  private serviceRpc?: RpcProvider;
+  protected serviceRpc?: RpcProvider;
 
   private vue: boolean;
