+++ test/integration/mocks/ApiIncrementalCheckerWithRpc.js
@@ -1,7 +1,7 @@
 const mock = require('mock-require');
 
 const origImport = require('../../../lib/ApiIncrementalChecker');
-const { rpcMethods, getRpcProvider } = require('../oldHelpers');
+const { rpcMethods, getRpcProvider } = require('../helpers/rpc');
 
 mock('../../../lib/ApiIncrementalChecker', {
   ApiIncrementalChecker: class extends origImport.ApiIncrementalChecker {
